backend/requirements.txt
# AI Wiki Quiz Generator - Backend Dependencies
# Python 3.10+

fastapi>=0.121.0
uvicorn[standard]>=0.38.0
sqlalchemy>=2.0.44
beautifulsoup4>=4.14.2
requests>=2.32.5
pydantic>=2.12.4
langchain-core>=1.0.3
langchain-google-genai>=3.0.1
# Removed direct google-generativeai dependency to avoid version conflicts
# langchain-google-genai will pull in the correct version
python-dotenv>=1.2.1
pymysql>=1.0.2
psycopg2-binary>=2.9.7
slowapi>=0.1.8